CONTAINS DETAILS Start implementing sending MIDI packet
Added write() function which, similar to USBMIDI, takes MIDIMessage as
parameter (which is nice, no need to deal with MIDI encodings ourselves)
and adds it to a buffer to be sent, conditionally sending the buffer
first if there is not enough space for the message.

The idea I have here is that upon initialisation of an RTPMIDI object, a
thread is setup to periodically perform the synchronization handshake,
and also periodically send and clear the midi buffer (if not empty).

Then the entire app is roughly as follows: sleep until button press,
then send NoteOn message. Whilst button pressed, poll accelerometer and
send PitchBend message based on accelerometer data. Once button is
released, go back to sleep.